refactor: reorganize types into a separate file
